done advanced task

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-income = float(input("Enter your monthly income: "))
-expenses = float(input("Enter your total monthly expenses: "))
+income = int(input("Enter your monthly income: "))
+expenses = int(input("Enter your total monthly expenses: "))
 
 savings = income - expenses
 projected_savings = savings * 12 + (savings * 12 * 0.05)
